requirements.txt
bs4==0.0.2 
openai==1.98.0
requests==2.32.4
python-dotenv==1.1.1
langfuse==3.2.4
Flask==3.1.1